requirements.txt
Django==1.9.3
MySQL-python==1.2.5
beautifulsoup4==4.4.1
django-filter==0.11.0
djangorestframework==3.3.2
icalendar==3.9.2
python-dateutil==2.5.0
pytz==2015.7
six==1.10.0
wsgiref==0.1.2
